worker: refactor: split nar_import along prefetch, relay, import and compression seams
